backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
google-generativeai==0.8.3
requests==2.32.4
pydantic==2.11.7
supabase==2.8.0
PyJWT==2.8.0
email-validator==2.1.0 